// src/service/model/prize/PrizeDetailDto.ts
/**
 * 抽奖API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { utils } from '../../utils';

/**
 * 奖项
 */
export class PrizeDetailDto {
  id?: number = undefined;

  /**
     * 奖项名
     */
  prizeItem?: string = undefined;

  /**
     * 商品名
     */
  prizeProduct?: string = undefined;

  /**
     * 价格
     */
  price?: string = undefined;

  /**
     * 总数量
     */
  quantity?: string | number = undefined;

  /**
     * 图片
     */
  img?: string = undefined;

  prizeId?: number = undefined;

  /**
     * 中奖率
     */
  pct?: number = undefined;

  staticImg?: any = undefined;

  static Format(data = {}): PrizeDetailDto {
    let obj: PrizeDetailDto = new PrizeDetailDto();
    obj['id'] = utils.convert(data['id'] || obj['id'], 'number');
    obj['prizeItem'] = utils.convert(
      data['prizeItem'] || obj['prizeItem'],
      'string'
    );
    obj['prizeProduct'] = utils.convert(
      data['prizeProduct'] || obj['prizeProduct'],
      'string'
    );
    obj['price'] = utils.convert(data['price'] || obj['price'], 'string');
    obj['quantity'] = utils.convert(
      data['quantity'] || obj['quantity'],
      'string'
    );
    obj['img'] = utils.convert(data['img'] || obj['img'], 'string');
    obj['prizeId'] = utils.convert(data['prizeId'] || obj['prizeId'], 'number');
    obj['pct'] = utils.convert(data['pct'] || obj['pct'], 'number');
    return obj;
  }
}